SlugFails and emptyBody to helper func
